requirements.txt
# requirements.txt

# Core LangChain Packages
langchain
langchain-core
langchain-community

# LLM Provider Packages
langchain-openai
langchain-google-genai
langchain-anthropic

# Vector Store and Document Loaders
chromadb
pypdf
python-docx

# Utilities
python-dotenv
pydantic
gradio